Created axiosCalls.js which helps make useEffect calls in App.jsx tinier
